chore: organize unit tests in pipelines package
