+++ packages/pipelines/test/unit/commands/disable.unit.test.ts
