+++ packages/pipelines/test/unit/commands/pipelines/destroy.unit.test.ts
