+++ packages/pipelines/test/unit/commands/pipelines/info.unit.test.ts
