+++ packages/pipelines/test/unit/commands/pipelines/open.unit.test.ts
