+++ packages/pipelines/test/unit/commands/pipelines/remove.unit.test.ts
