+++ packages/pipelines/test/unit/commands/pipelines/rename.unit.test.ts
