+++ packages/pipelines/test/unit/commands/pipelines/setup.unit.test.ts
@@ -2,7 +2,7 @@ import {CliUx} from '@oclif/core'
 import color from '@heroku-cli/color'
 import {expect, test} from '@oclif/test'
 import sinon from 'sinon'
-import pollAppSetups from '../../../src/setup/poll-app-setups'
+import pollAppSetups from '../../../../src/setup/poll-app-setups'
 
 const cli = CliUx.ux
